This might be better maybe dunno hopefully
